+++ CHANGELOG.md
@@ -4,6 +4,29 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This change log adheres to standards from [Keep a CHANGELOG](http://keepachangelog.com).
 
 ## [Unreleased]
+### Added
+- `max-lines` new rule from eslint v2.12.0
+- `no-mixed-operators` new rule from eslint v2.12.0
+- `no-prototype-builtins` new rule from eslint v2.11.0
+- `no-unsafe-finally` new rule from eslint v2.9.0
+- `no-useless-computed-key` new rule from eslint v2.9.0
+- `no-useless-rename` new rule from eslint v2.11.0
+- `object-curly-newline` new rule from eslint v2.12.0
+- `object-property-newline` new rule from eslint v2.10.0
+- `rest-spread-spacing` new rule from eslint v2.12.0
+- `unicode-bom` new rule from eslint v2.11.0
+- `react/jsx-filename-extension` new rule from eslint-plugin-react v5.2.0
+- `react/jsx-no-target-blank` new rule from eslint-plugin-react v5.1.0
+- `react/no-comment-textnodes` new rule from eslint-plugin-react v5.2.0
+- `react/no-render-return-value` new rule from eslint-plugin-react v5.2.0
+- `react/require-optimization` new rule from eslint-plugin-react v5.2.0
+
+### Changed
+- Amended documentation
+- Explicited default values for some rules (without value modification)
+- `array-bracket-spacing` option changed to always
+- `object-curly-spacing` option changed to always
+- `react/jsx-space-before-closing` option changed to always